Revert "fixed formatting"
This reverts commit 8e851e93e04e8058b15936f9276db635f62d7aee.

+++ src/main/java/at/pavlov/cannons/cannon/DesignStorage.java
@@ -222,7 +222,7 @@ public class DesignStorage
         cannonDesign.setMassOfCannon(cannonDesignConfig.getInt("realisticBehaviour.massOfCannon", 1000));//What means 1000?
         
         // overloading stuff
-        cannonDesign.setOverloadingType(OverloadingType.get(cannonDesignConfig.getInt("realisticBehaviour.Overloading.Mode")));
+        cannonDesign.setOverloadingType(OverloadingType.get((byte) (cannonDesignConfig.getInt("realisticBehaviour.Overloading.Mode"))));
         cannonDesign.setOverloadingExponent(cannonDesignConfig.getDouble("realisticBehaviour.Overloading.Exponent"));
         cannonDesign.setOverloadingChangeInc((cannonDesignConfig.getDouble("realisticBehaviour.Overloading.ChanceInc")));
         cannonDesign.setOverloadingMaxOverloadableGunpowder((cannonDesignConfig.getInt("realisticBehaviour.Overloading.MaxOverloadableGunpowder")));
